fin libft.h

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: saneveu <saneveu@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 17:52:49 by saneveu           #+#    #+#             */
-/*   Updated: 2018/11/14 02:11:33 by saneveu          ###   ########.fr       */
+/*   Updated: 2018/11/14 20:49:18 by saneveu          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *str)
+size_t		ft_strlen(const char *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str[i])
